Refactor: seperate Logout component
